requirements.txt
# Dependências do projeto

fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
strands-agents>=1.25.0
strands-agents-tools
pydantic==2.9.0
ollama